// src/modules/buyer/redux/order.types.ts
export enum OrderTypes {
  ORDER_ACTION_START = "ORDER_ACTION_START",
  DELIVERY_ACTION_START = "DELIVERY_ACTION_START",

  ORDER_GET_SUCCESS = "ORDER_GET_SUCCESS",
  DELIVERY_GET_SUCCESS = "DELIVERY_GET_SUCCESS",

  ORDER_GET_FAILED = "ORDER_GET_FAILED",
  DELIVERY_GET_FAILED = "DELIVERY_GET_FAILED",

  ORDER_CREATE_SUCCESS = "ORDER_CREATE_SUCCESS",
  ORDER_CREATE_FAILED = "ORDER_CREATE_FAILED",
  ORDER_GET_BY_ID_SUCCESS = "ORDER_GET_BY_ID_SUCCESS",
  ORDER_GET_BY_ID_FAILED = "ORDER_GET_BY_ID_FAILED",
  ORDER_UPDATE_SUCCESS = "ORDER_UPDATE_SUCCESS",
  ORDER_DELETE_SUCCESS = "ORDER_DELETE_SUCCESS",
  ORDER_DELETE_FAILED = "ORDER_DELETE_FAILED",
  ORDER_UPDATE_FAILED = "ORDER_UPDATE_FAILED",

  DELIVERY_ACTION_END = "DELIVERY_ACTION_END",
  ORDER_ACTION_END = "ORDER_ACTION_END",
}